Split news post tasks
