+++ handlers/news/newsEditPostTask.go
@@ -1,0 +1,72 @@
+package news
+
+import (
+	"database/sql"
+	"github.com/arran4/goa4web/core/consts"
+	"net/http"
+	"net/url"
+	"strconv"
+
+	"github.com/gorilla/mux"
+
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/handlers"
+	"github.com/arran4/goa4web/internal/db"
+	notif "github.com/arran4/goa4web/internal/notifications"
+	"github.com/arran4/goa4web/internal/tasks"
+)
+
+type EditTask struct{ tasks.TaskString }
+
+var editTask = &EditTask{TaskString: TaskEdit}
+
+var _ tasks.Task = (*EditTask)(nil)
+var _ notif.AdminEmailTemplateProvider = (*EditTask)(nil)
+
+func (EditTask) AdminEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("adminNotificationNewsEditEmail")
+}
+
+func (EditTask) AdminInternalNotificationTemplate() *string {
+	v := notif.NotificationTemplateFilenameGenerator("adminNotificationNewsEditEmail")
+	return &v
+}
+
+func (EditTask) Action(w http.ResponseWriter, r *http.Request) any {
+	if err := handlers.ValidateForm(r, []string{"language", "text"}, []string{"language", "text"}); err != nil {
+		r.URL.RawQuery = "error=" + url.QueryEscape(err.Error())
+		handlers.TaskErrorAcknowledgementPage(w, r)
+		return nil
+	}
+	languageId, err := strconv.Atoi(r.PostFormValue("language"))
+	if err != nil {
+		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
+		return nil
+	}
+	text := r.PostFormValue("text")
+	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+	vars := mux.Vars(r)
+	postId, _ := strconv.Atoi(vars["post"])
+
+	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
+	if !cd.HasGrant("news", "post", "edit", int32(postId)) {
+		r.URL.RawQuery = "error=" + url.QueryEscape("Forbidden")
+		handlers.TaskErrorAcknowledgementPage(w, r)
+		return nil
+	}
+	err = queries.UpdateNewsPost(r.Context(), db.UpdateNewsPostParams{
+		Idsitenews:         int32(postId),
+		LanguageIdlanguage: int32(languageId),
+		News: sql.NullString{
+			String: text,
+			Valid:  true,
+		},
+	})
+	if err != nil {
+		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
+		return nil
+	}
+
+	handlers.TaskDoneAutoRefreshPage(w, r)
+	return nil
+}
