+++ handlers/news/newsPostPage.go
@@ -7,7 +7,6 @@ import (
 	"github.com/arran4/goa4web/core/consts"
 	"log"
 	"net/http"
-	"net/url"
 	"strconv"
 
 	"github.com/gorilla/mux"
@@ -19,11 +18,6 @@ import (
 	"github.com/arran4/goa4web/core/templates"
 	"github.com/arran4/goa4web/handlers"
 	"github.com/arran4/goa4web/internal/db"
-	"github.com/arran4/goa4web/internal/eventbus"
-	notif "github.com/arran4/goa4web/internal/notifications"
-	"github.com/arran4/goa4web/internal/tasks"
-	"github.com/arran4/goa4web/workers/postcountworker"
-	"github.com/arran4/goa4web/workers/searchworker"
 )
 
 type NewsPost struct {
@@ -31,123 +25,6 @@ type NewsPost struct {
 	// ShowEdit is true when the current user can modify the post. Users with
 	// the writer, moderator or administrator role are permitted to edit.
 	ShowEdit bool
-}
-
-type ReplyTask struct{ tasks.TaskString }
-
-// ReplyTask sends notifications and auto-subscribes authors and followers when
-// someone replies to a news post.
-var (
-	replyTask = &ReplyTask{TaskString: TaskReply}
-
-	_ tasks.Task                                    = (*ReplyTask)(nil)
-	_ notif.SubscribersNotificationTemplateProvider = (*ReplyTask)(nil)
-	_ notif.AdminEmailTemplateProvider              = (*ReplyTask)(nil)
-	_ notif.AutoSubscribeProvider                   = (*ReplyTask)(nil)
-)
-
-func (ReplyTask) IndexType() string { return searchworker.TypeComment }
-
-func (ReplyTask) IndexData(data map[string]any) []searchworker.IndexEventData {
-	if v, ok := data[searchworker.EventKey].(searchworker.IndexEventData); ok {
-		return []searchworker.IndexEventData{v}
-	}
-	return nil
-}
-
-var _ searchworker.IndexedTask = ReplyTask{}
-
-func (ReplyTask) SubscribedEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("replyEmail")
-}
-
-func (ReplyTask) SubscribedInternalNotificationTemplate() *string {
-	s := notif.NotificationTemplateFilenameGenerator("reply")
-	return &s
-}
-
-func (ReplyTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationNewsReplyEmail")
-}
-
-func (ReplyTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationNewsReplyEmail")
-	return &v
-}
-
-// AutoSubscribePath registers this reply so the author automatically follows
-// subsequent comments on the news post.
-// When users reply to a news post we automatically subscribe them so
-// they receive updates to the thread they just engaged with.
-// AutoSubscribePath allows commenters to automatically watch for further replies.
-// AutoSubscribePath implements notif.AutoSubscribeProvider. A subscription to
-// the underlying discussion thread is created using event data when available.
-func (ReplyTask) AutoSubscribePath(evt eventbus.TaskEvent) (string, string, error) {
-	if data, ok := evt.Data[postcountworker.EventKey].(postcountworker.UpdateEventData); ok {
-		return string(TaskReply), fmt.Sprintf("/forum/topic/%d/thread/%d", data.TopicID, data.ThreadID), nil
-	}
-	return string(TaskReply), evt.Path, nil
-}
-
-type EditTask struct{ tasks.TaskString }
-
-var editTask = &EditTask{TaskString: TaskEdit}
-
-var _ tasks.Task = (*EditTask)(nil)
-var _ notif.AdminEmailTemplateProvider = (*EditTask)(nil)
-
-func (EditTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationNewsEditEmail")
-}
-
-func (EditTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationNewsEditEmail")
-	return &v
-}
-
-type NewPostTask struct{ tasks.TaskString }
-
-var newPostTask = &NewPostTask{TaskString: TaskNewPost}
-
-// NewPostTask sends notifications when a new post is created and automatically
-// subscribes the author to future replies.
-var (
-	_ tasks.Task                                    = (*NewPostTask)(nil)
-	_ notif.SubscribersNotificationTemplateProvider = (*NewPostTask)(nil)
-	_ notif.AdminEmailTemplateProvider              = (*NewPostTask)(nil)
-	_ notif.AutoSubscribeProvider                   = (*NewPostTask)(nil)
-)
-
-func (NewPostTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationNewsAddEmail")
-}
-
-func (NewPostTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationNewsAddEmail")
-	return &v
-}
-
-func (NewPostTask) SubscribedEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("newsAddEmail")
-}
-
-func (NewPostTask) SubscribedInternalNotificationTemplate() *string {
-	s := notif.NotificationTemplateFilenameGenerator("news_add")
-	return &s
-}
-
-// AutoSubscribePath links the newly created post so that any future replies
-// notify the author by default.
-// Subscribing the poster ensures they are notified when readers engage
-// with their new thread.
-// AutoSubscribePath keeps authors in the loop on new post discussions.
-// AutoSubscribePath implements notif.AutoSubscribeProvider. Subscriptions use
-// the thread path derived from postcountworker data when possible.
-func (NewPostTask) AutoSubscribePath(evt eventbus.TaskEvent) (string, string, error) {
-	if data, ok := evt.Data[postcountworker.EventKey].(postcountworker.UpdateEventData); ok {
-		return string(TaskNewPost), fmt.Sprintf("/forum/topic/%d/thread/%d", data.TopicID, data.ThreadID), nil
-	}
-	return string(TaskNewPost), evt.Path, nil
 }
 
 func NewsPostPage(w http.ResponseWriter, r *http.Request) {
@@ -319,246 +196,4 @@ func NewsPostPage(w http.ResponseWriter, r *http.Request) {
 	}
 
 	handlers.TemplateHandler(w, r, "postPage.gohtml", data)
-}
-
-func (ReplyTask) Action(w http.ResponseWriter, r *http.Request) any {
-	session, ok := core.GetSessionOrFail(w, r)
-	if !ok {
-		return nil
-	}
-
-	vars := mux.Vars(r)
-	pid, err := strconv.Atoi(vars["post"])
-
-	if err != nil {
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	}
-	if pid == 0 {
-		log.Printf("Error: no bid")
-		http.Redirect(w, r, "?error="+"No bid", http.StatusTemporaryRedirect)
-		return nil
-	}
-
-	uid, _ := session.Values["UID"].(int32)
-
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-	if !cd.HasGrant("news", "post", "reply", int32(pid)) {
-		http.Error(w, "Forbidden", http.StatusForbidden)
-		return nil
-	}
-
-	post, err := queries.GetNewsPostByIdWithWriterIdAndThreadCommentCount(r.Context(), db.GetNewsPostByIdWithWriterIdAndThreadCommentCountParams{
-		ViewerID: uid,
-		ID:       int32(pid),
-		UserID:   sql.NullInt32{Int32: uid, Valid: uid != 0},
-	})
-	if err != nil {
-		log.Printf("GetNewsPostByIdWithWriterIdAndThreadCommentCountForUser Error: %s", err)
-		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-		return nil
-	}
-
-	var pthid = post.ForumthreadID
-	pt, err := queries.FindForumTopicByTitle(r.Context(), sql.NullString{
-		String: NewsTopicName,
-		Valid:  true,
-	})
-	var ptid int32
-	if errors.Is(err, sql.ErrNoRows) {
-		ptidi, err := queries.CreateForumTopic(r.Context(), db.CreateForumTopicParams{
-			ForumcategoryIdforumcategory: 0,
-			Title: sql.NullString{
-				String: NewsTopicName,
-				Valid:  true,
-			},
-			Description: sql.NullString{
-				String: NewsTopicDescription,
-				Valid:  true,
-			},
-		})
-		if err != nil {
-			log.Printf("Error: createForumTopic: %s", err)
-			http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-			return nil
-		}
-		ptid = int32(ptidi)
-	} else if err != nil {
-		log.Printf("Error: findForumTopicByTitle: %s", err)
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	} else {
-		ptid = pt.Idforumtopic
-	}
-	if pthid == 0 {
-		pthidi, err := queries.MakeThread(r.Context(), ptid)
-		if err != nil {
-			log.Printf("Error: makeThread: %s", err)
-			http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-			return nil
-		}
-		pthid = int32(pthidi)
-		if err := queries.AssignNewsThisThreadId(r.Context(), db.AssignNewsThisThreadIdParams{
-			ForumthreadID: pthid,
-			Idsitenews:    int32(pid),
-		}); err != nil {
-			log.Printf("Error: assign_news_to_thread: %s", err)
-			http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-			return nil
-		}
-	}
-
-	text := r.PostFormValue("replytext")
-	languageId, _ := strconv.Atoi(r.PostFormValue("language"))
-
-	endUrl := cd.AbsoluteURL(fmt.Sprintf("/news/news/%d", pid))
-
-	evt := cd.Event()
-	if evt.Data == nil {
-		evt.Data = map[string]any{}
-	}
-	evt.Data["CommentURL"] = endUrl
-
-	cid, err := queries.CreateComment(r.Context(), db.CreateCommentParams{
-		LanguageIdlanguage: int32(languageId),
-		UsersIdusers:       uid,
-		ForumthreadID:      pthid,
-		Text: sql.NullString{
-			String: text,
-			Valid:  true,
-		},
-	})
-	if err != nil {
-		log.Printf("Error: createComment: %s", err)
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	}
-
-	if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-		if evt := cd.Event(); evt != nil {
-			if evt.Data == nil {
-				evt.Data = map[string]any{}
-			}
-			evt.Data[postcountworker.EventKey] = postcountworker.UpdateEventData{ThreadID: pthid, TopicID: ptid}
-		}
-	}
-	if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-		if evt := cd.Event(); evt != nil {
-			if evt.Data == nil {
-				evt.Data = map[string]any{}
-			}
-			evt.Data[searchworker.EventKey] = searchworker.IndexEventData{Type: searchworker.TypeComment, ID: int32(cid), Text: text}
-		}
-	}
-
-	handlers.TaskDoneAutoRefreshPage(w, r)
-	return nil
-}
-
-func (EditTask) Action(w http.ResponseWriter, r *http.Request) any {
-	if err := handlers.ValidateForm(r, []string{"language", "text"}, []string{"language", "text"}); err != nil {
-		r.URL.RawQuery = "error=" + url.QueryEscape(err.Error())
-		handlers.TaskErrorAcknowledgementPage(w, r)
-		return nil
-	}
-	languageId, err := strconv.Atoi(r.PostFormValue("language"))
-	if err != nil {
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	}
-	text := r.PostFormValue("text")
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	vars := mux.Vars(r)
-	postId, _ := strconv.Atoi(vars["post"])
-
-	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-	if !cd.HasGrant("news", "post", "edit", int32(postId)) {
-		r.URL.RawQuery = "error=" + url.QueryEscape("Forbidden")
-		handlers.TaskErrorAcknowledgementPage(w, r)
-		return nil
-	}
-	err = queries.UpdateNewsPost(r.Context(), db.UpdateNewsPostParams{
-		Idsitenews:         int32(postId),
-		LanguageIdlanguage: int32(languageId),
-		News: sql.NullString{
-			String: text,
-			Valid:  true,
-		},
-	})
-	if err != nil {
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	}
-
-	handlers.TaskDoneAutoRefreshPage(w, r)
-	return nil
-}
-
-func (NewPostTask) Action(w http.ResponseWriter, r *http.Request) any {
-	if err := handlers.ValidateForm(r, []string{"language", "text"}, []string{"language", "text"}); err != nil {
-		r.URL.RawQuery = "error=" + url.QueryEscape(err.Error())
-		handlers.TaskErrorAcknowledgementPage(w, r)
-		return nil
-	}
-	languageId, err := strconv.Atoi(r.PostFormValue("language"))
-	if err != nil {
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	}
-	text := r.PostFormValue("text")
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	session, ok := core.GetSessionOrFail(w, r)
-	if !ok {
-		return nil
-	}
-	uid, _ := session.Values["UID"].(int32)
-
-	if cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData); !cd.HasGrant("news", "post", "post", 0) {
-		r.URL.RawQuery = "error=" + url.QueryEscape("Forbidden")
-		handlers.TaskErrorAcknowledgementPage(w, r)
-		return nil
-	}
-	id, err := queries.CreateNewsPost(r.Context(), db.CreateNewsPostParams{
-		LanguageIdlanguage: int32(languageId),
-		News: sql.NullString{
-			String: text,
-			Valid:  true,
-		},
-		UsersIdusers: uid,
-	})
-	if err != nil {
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	}
-
-	// give the author edit rights on the new post
-	if _, err := queries.CreateGrant(r.Context(), db.CreateGrantParams{
-		UserID:   sql.NullInt32{Int32: uid, Valid: true},
-		RoleID:   sql.NullInt32{},
-		Section:  "news",
-		Item:     sql.NullString{String: "post", Valid: true},
-		RuleType: "allow",
-		ItemID:   sql.NullInt32{Int32: int32(id), Valid: true},
-		ItemRule: sql.NullString{},
-		Action:   "edit",
-		Extra:    sql.NullString{},
-	}); err != nil {
-		log.Printf("create grant: %v", err)
-	}
-
-	if u, err := queries.GetUserById(r.Context(), uid); err == nil {
-		if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-			if evt := cd.Event(); evt != nil {
-				if evt.Data == nil {
-					evt.Data = map[string]any{}
-				}
-				evt.Data["Author"] = u.Username.String
-				evt.Data["PostURL"] = cd.AbsoluteURL(fmt.Sprintf("/news/news/%d", id))
-			}
-		}
-	}
-
-	handlers.TaskDoneAutoRefreshPage(w, r)
-	return nil
 }
